pyproject.toml
[project]
name = "kycc"
version = "0.1.0"
description = "Know Your Coin History — Bitcoin transaction graph explorer"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "python-bitcoinrpc>=1.0",
    "python-bitcoinlib>=0.12",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
